fix zshenv so that tramp is happy
emacs tramp-mode uses ssh to open files locally and remotely. It
needs a simple prompt, and will fail if there is a fancy one. I do
have that, so I have to make sure that when I connect using tramp,
it will revert to a simple one. As tramp sets the $TERM environment
variable to 'dumb', all we have to do is to make sure the fancy
stuff is disabled when that variable is set to that value.